--- backend/challenge-service/src/main/java/com/itachallenges/challengeservice/submission/application/dto/SaveDraftSubmissionCommand.java
@@ -1,7 +1,0 @@
-package com.itachallenges.challengeservice.submission.application.dto;
-
-public record SaveDraftSubmissionCommand(
-        String challengeId,
-        String userId,
-        String code
-) {}
